Updated `lofty` to 0.23 and `rand` to 0.10 and fixed compatibility errors

+++ src/library/album.rs
@@ -9,7 +9,7 @@ use crate::player::queue_item::QueueItem;
 
 pub struct Album {
     pub title: String,
-    pub year: u32,
+    pub year: u16,
     pub songs: AlbumSongs,
     pub artist: SharedArtist,
 }
